Fix linter and ESlint errors
