+++ src/App.css
@@ -30,10 +30,12 @@
   0% {
     transform: translateY(0);
   }
+
   50% {
     transform: translateY(10px);
   }
+
   100% {
-    transform: translateY(0px);
+    transform: translateY(0);
   }
 }
