+++ src/App.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import MyReservations from './components/Myreservations';
 import { Route, Routes } from 'react-router-dom';
 import Event from './components/events/Event';
 import Navbar from './components/navbar/Navbar';
@@ -10,8 +9,6 @@ import './App.css';
 function App() {
   return (
     <div className="App">
-      <p>Reservations list</p>
-      <MyReservations />
       <Navbar />
       <Routes>
         <Route path="/" element={<Event />} />
