+++ src/components/events/Event.js
@@ -5,7 +5,6 @@ import { fetchEvents, selectEvents } from '../../features/events/eventSlice';
 export default function Event() {
   const dispatch = useDispatch();
   const events = useSelector(selectEvents);
-  console.log(events);
   useEffect(() => {
     dispatch(fetchEvents());
   }, []);
