+++ src/components/navbar/sidebar.css
@@ -1,54 +1,54 @@
 @import url('https://fonts.googleapis.com/css2?family=Dancing+Script&display=swap');
 
 .sidebar {
-    position: fixed;
-    top: 0;
-    left: 0;
-    height: 100vh;
-    width: 260px;
-    background-color: #fff;
-    border-right: 1px solid #EDEDED;
+  position: fixed;
+  top: 0;
+  left: 0;
+  height: 100vh;
+  width: 260px;
+  background-color: #fff;
+  border-right: 1px solid #ededed;
 }
 
 .sidebar__logo {
-    display: grid;
-    place-items: center;
-    height: 120px;
-    font-size: 1.5rem;
-    font-weight: 700;
-    font-family: 'Dancing Script', cursive;
+  display: grid;
+  place-items: center;
+  height: 120px;
+  font-size: 1.5rem;
+  font-weight: 700;
+  font-family: 'Dancing Script', cursive;
 }
+
 .sidebar__menu {
-    position: relative;
+  position: relative;
 }
 
 .sidebar__menu__item {
-    display: flex;
-    align-items: center;
-    place-content: flex-start;
-    padding: 1rem 3rem;
-    font-size: 1.1rem;
-    font-weight: 500;
-    color: black;
+  display: flex;
+  align-items: center;
+  place-content: flex-start;
+  padding: 1rem 3rem;
+  font-size: 1.1rem;
+  font-weight: 500;
+  color: black;
 }
 
 .sidebar__menu__item__icon {
-    margin-right: 1rem;
+  margin-right: 1rem;
 }
 
 .pageLink {
-    text-decoration: none;
+  text-decoration: none;
 }
 
-
 .sidebar__menu__indicator {
-    position: absolute;
-    top: 0;
-    left: 50%;
-    width: calc(100% - 40px);
-    border-radius: 10px;
-    background-color: #308efe;
-    z-index: -1;
-    transform: translateX(-50%);
-    transition: 0.3s ease-in-out;
+  position: absolute;
+  top: 0;
+  left: 50%;
+  width: calc(100% - 40px);
+  border-radius: 10px;
+  background-color: #308efe;
+  z-index: -1;
+  transform: translateX(-50%);
+  transition: 0.3s ease-in-out;
 }
